convert to use datastore
